botlemos/requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.4.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
future==0.18.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.4
python-dotenv==0.10.3
python-telegram-bot==12.2.0
six==1.13.0
telepot==12.7
tornado==6.0.3
urllib3==1.25.7
wrapt==1.11.2
yarl==1.4.2
